server/python/requirements.txt
pymupdf>=1.24.0
rapidocr-onnxruntime>=1.4.3
Pillow>=10.0.0
numpy>=1.26.0
